(bifrost, ebysus) [REF] move some things from EbysusData to BifrostData. [REF] use get_param() to 'cleanup' some code for BifrostData. [REF] fix one possible source of memory leak in bifrost.py (weakref to obj for Cross_sect), however the memory leak seems to still exist.
