+++ helita/sim/bifrost.py
@@ -5,6 +5,7 @@ Set of programs to read and interact with output from Bifrost
 # import builtin modules
 import os
 import functools
+import weakref
 from glob import glob
 
 # import external public modules
@@ -136,8 +137,7 @@ class BifrostData(object):
         self.transunits = False
         self.cross_sect = cross_sect_for_obj(self)
         if 'tabinputfile' in self.params.keys(): 
-            tabfile = os.path.join(self.fdir, 
-                self.params['tabinputfile'][self.snapInd].strip())
+            tabfile = os.path.join(self.fdir, self.get_param('tabinputfile').strip())
             if os.access(tabfile, os.R_OK):
                 self.rhoee = Rhoeetab(tabfile=tabfile, fdir=fdir, radtab=True)
 
@@ -149,20 +149,18 @@ class BifrostData(object):
             Sets list of avaible variables
         """
         self.snapvars = ['r', 'px', 'py', 'pz', 'e']
-        self.auxvars = self.params['aux'][self.snapInd].split()
+        self.auxvars = self.get_param('aux', error_prop=True).split()
         if self.do_mhd:
             self.snapvars += ['bx', 'by', 'bz']
         self.hionvars = []
         self.heliumvars = []
-        if 'do_hion' in self.params:
-            if self.params['do_hion'][self.snapInd] > 0:
-                self.hionvars = ['hionne', 'hiontg', 'n1',
-                                 'n2', 'n3', 'n4', 'n5', 'n6', 'nh2']
-                self.hion = True
-        if 'do_helium' in self.params:
-            if self.params['do_helium'][self.snapInd] > 0:
-                self.heliumvars = ['nhe1', 'nhe2', 'nhe3']
-                self.heion = True
+        if self.get_param('do_hion', default=0) > 0:
+            self.hionvars = ['hionne', 'hiontg', 'n1',
+                             'n2', 'n3', 'n4', 'n5', 'n6', 'nh2']
+            self.hion = True
+        if self.get_param('do_helium', default=0) > 0:
+            self.heliumvars = ['nhe1', 'nhe2', 'nhe3']
+            self.heion = True
         self.compvars = ['ux', 'uy', 'uz', 's', 'ee']
         self.simple_vars = self.snapvars + self.auxvars + self.hionvars + \
             self.heliumvars
@@ -303,6 +301,34 @@ class BifrostData(object):
         self.time = self.params['t']
         if self.sel_units=='cgs': 
             self.time *= self.uni.uni['t']
+
+    def get_param(self, param, default=None, warning=None, error_prop=None):
+        ''' get param via self.params[param][self.snapInd].
+
+        if param not in self.params.keys(), then the following kwargs may play a role:
+            default: None (default) or any value.
+                return this value (eventually) instead. (check warning and error_prop first.)
+            warning: None (default) or any Warning or string.
+                if not None, do warnings.warn(warning).
+            error_prop: None (default), True, or any Exception object.
+                None --> ignore this kwarg.
+                True --> raise the original KeyError caused by trying to get self.params[param].
+                else --> raise error_prop from None.
+        '''
+        try:
+            p = self.params[param]
+        except KeyError as err_triggered:
+            if warning is not None:
+                warnings.warn(warning)
+            if error_prop is not None:
+                if isinstance(error_prop, BaseException):
+                    raise error_prop from None  # "from None" --> show just this error, not also err_triggered
+                elif error_prop:
+                    raise err_triggered
+            return default
+        else:
+            p = p[self.snapInd]
+        return p
         
     def __read_mesh(self, meshfile, firstime=False):
         """
@@ -310,7 +336,7 @@ class BifrostData(object):
         """
         if meshfile is None:
             meshfile = os.path.join(
-                self.fdir, self.params['meshfile'][self.snapInd].strip())
+                self.fdir, self.get_param('meshfile', error_prop=True).strip())
         if os.path.isfile(meshfile):
             f = open(meshfile, 'r')
             for p in ['x', 'y', 'z']:
@@ -849,7 +875,7 @@ class BifrostData(object):
             filename += fsuffix_a + fsuffix_b
         elif var in self.hionvars:
             idx = self.hionvars.index(var)
-            isnap = self.params['isnap'][self.snapInd]
+            isnap = self.get_param('isnap', error_prop=True)
             if panic: 
                 filename = filename + '.hion.panic'
             else: 
@@ -863,7 +889,7 @@ class BifrostData(object):
                         filename = '%s_.hion%s.snap' % (self.file_root, isnap)
         elif var in self.heliumvars:
             idx = self.heliumvars.index(var)
-            isnap = self.params['isnap'][self.snapInd]
+            isnap = self.get_param('isnap', error_prop=True)
             if panic: 
                 filename = filename + '.helium.panic'
             else:
@@ -965,7 +991,7 @@ class BifrostData(object):
         else:
             rho = self.r[sx, sy, sz] * self.uni.u_r
             subsfile = os.path.join(self.fdir, 'subs.dat')
-            tabfile = os.path.join(self.fdir, self.params['tabinputfile'][self.snapInd].strip())
+            tabfile = os.path.join(self.fdir, self.get_param('tabinputfile', error_prop=True).strip())
             tabparams = []
             if os.access(tabfile, os.R_OK):
                 tabparams = read_idl_ascii(tabfile, obj=self)
@@ -1138,6 +1164,55 @@ class BifrostData(object):
             fout2.write("\n%i\n" % nz)
             z.tofile(fout2, sep="  ", format="%11.5e")
             fout2.close()
+
+    # --- misc. convenience methods --- #
+
+    def get_varm(self, *args__get_var, **kwargs__get_var):
+        '''get_var but returns np.mean() of result.
+        provided for convenience for quicker debugging.
+        '''
+        return np.mean(self.get_var(*args__get_var, **kwargs__get_var))
+
+    def zero(self):
+        '''return np.zeros_like(self.r, subok=False).
+        (an array of zeros with same shape and dtype like self.r but which is not a memmap.)
+        '''
+        return np.zeros_like(self.r, subok=False)
+
+    def get_lmin(self):
+        '''return smallest length resolvable for each direction ['x', 'y', 'z'].
+        result is in [simu. length units]. Multiply by self.uni.usi_l to convert to SI.
+
+        return 1 (instead of 0) for any direction with number of points < 2.
+        '''
+        def _dxmin(x):
+            dx1d = getattr(self, 'd'+x+'1d')
+            if len(dx1d)==1:
+                return 1
+            else:
+                return dx1d.min()
+        return np.array([_dxmin(x) for x in AXES])
+
+    def get_kmax(self):
+        '''return largest value of each component of wavevector resolvable by self.
+        I.e. returns [max kx, max ky, max kz].
+        result is in [1/ simu. length units]. Divide by self.uni.usi_l to convert to SI.
+        '''
+        return 2 * np.pi / self.get_lmin()
+
+    def get_unit_vector(self, var, mean=False, **kw__get_var):
+        '''return unit vector of var. [varx, vary, varz]/|var|.'''
+        varx = self.get_var(var+'x', **kw__get_var)
+        vary = self.get_var(var+'y', **kw__get_var)
+        varz = self.get_var(var+'z', **kw__get_var)
+        varmag = self.get_var('mod'+var, **kw__get_var)
+        if mean:
+            varx, vary, varz, varmag = varx.mean(), vary.mean(), varz.mean(), varmag.mean()
+        return np.array([varx, vary, varz]) / varmag
+
+    if file_memory.DEBUG_MEMORY_LEAK:
+        def __del__(self):
+            print('deleted {}'.format(self), flush=True)
 
 
 def write_br_snap(rootname,r,px,py,pz,e,bx,by,bz):
@@ -2025,9 +2100,10 @@ class Cross_sect:
         if verbose is None: 
             verbose = False if obj is None else getattr(obj, 'verbose', False)
         self.verbose = verbose
-        self.obj = obj
         self.kelvin = kelvin
         self.units = {True: 'K', False: 'eV'}[self.kelvin]
+        # save pointer to obj. Use weakref to help ensure we don't create a circular reference.
+        self.obj = (lambda: None) if (obj is None) else weakref.ref(obj)  # self.obj() returns obj.
         # read table file and calculate parameters
         if cross_tab is None:
             cross_tab = ['h-h-data2.txt', 'h-h2-data.txt', 'he-he.txt',
@@ -2049,7 +2125,7 @@ class Cross_sect:
         self.cross_tab = dict()
         for itab in range(len(self.cross_tab_list)):
             self.cross_tab[itab] = read_cross_txt(self.cross_tab_list[itab],firstime=firstime,
-                                                  obj=getattr(self, 'obj', None), kelvin=self.kelvin)
+                                                  obj=self.obj(), kelvin=self.kelvin)
             
 
     def tab_interp(self, tg, itab=0, out='el', order=1):
@@ -2191,7 +2267,8 @@ def bifrost2d_to_rh15d(snaps, outfile, file_root, meshfile, fdir, writeB=False,
 @file_memory.remember_and_recall('_memory_read_idl_ascii')
 def read_idl_ascii(filename,firstime=False):
     ''' Reads IDL-formatted (command style) ascii file into dictionary.
-    if obj is not None,'''
+    if obj is not None, remember the result and restore it if ever reading the same exact file again.
+    '''
     li = 0
     params = {}
     # go through the file, add stuff to dictionary
